--- src/server/src/main/java/com/advt/server/config/WebConfig.java
@@ -1,4 +1,0 @@
-package com.advt.server.config;
-
-public class WebConfig {
-}
