--- src/server/src/main/java/com/advt/server/controller/TestController.java
@@ -1,4 +1,0 @@
-package com.advt.server.controller;
-
-public class TestController {
-}
