--- src/server/src/main/java/com/advt/server/controller/dto/request/TestRequest.java
@@ -1,4 +1,0 @@
-package com.advt.server.controller.dto.request;
-
-public class TestRequest {
-}
