--- src/server/src/main/java/com/advt/server/controller/dto/response/TestResponse.java
@@ -1,4 +1,0 @@
-package com.advt.server.controller.dto.response;
-
-public class TestResponse {
-}
