--- src/server/src/main/java/com/advt/server/service/TestService.java
@@ -1,4 +1,0 @@
-package com.advt.server.service;
-
-public class TestService {
-}
