+++ src/server/src/main/java/com/drm/server/ServerApplication.java
@@ -1,4 +1,4 @@
-package com.advt.server;
+package com.drm.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
