+++ src/server/src/main/java/com/drm/server/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.advt.server.config;
+package com.drm.server.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
