+++ src/server/src/main/java/com/drm/server/config/WebConfig.java
@@ -1,0 +1,4 @@
+package com.drm.server.config;
+
+public class WebConfig {
+}
