+++ src/server/src/main/java/com/drm/server/controller/TestController.java
@@ -1,0 +1,4 @@
+package com.drm.server.controller;
+
+public class TestController {
+}
