+++ src/server/src/main/java/com/drm/server/controller/dto/request/TestRequest.java
@@ -1,0 +1,4 @@
+package com.drm.server.controller.dto.request;
+
+public class TestRequest {
+}
