+++ src/server/src/main/java/com/drm/server/controller/dto/response/TestResponse.java
@@ -1,0 +1,4 @@
+package com.drm.server.controller.dto.response;
+
+public class TestResponse {
+}
