+++ src/server/src/main/java/com/drm/server/service/TestService.java
@@ -1,0 +1,4 @@
+package com.drm.server.service;
+
+public class TestService {
+}
