+++ src/server/src/test/java/com/drm/server/ServerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.advt.server;
+package com.drm.server;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
